// src/interfaces/vehicles.interface.ts
export interface IVehicle {
    id: number;
    license_plate_number: string;
    vehicle_model: string;
    year_of_manufacture: number;
    vin_identification_number: string;
    horse_power: number;
    capacity: number;
    load_capacity: number;
    insurance_information: string;
    tour: number;
};